feat: loads of changes, short summary below:
* Now subscribes to the scraper web api to get departures from stops.
* Created ThemeContext for holding the line colors
  and maybe dark mode in the future

Note to self: Need to add special check if a search results in trips on two different dates (in scraper).
Currently unable to see what direction a a line is going in.
Need to figure out a good way of doing this.
Maybe regarding each direction of a line as two seperate lines.